Extract path_helpers.rs from turn.rs (parent #680)
Create new sibling module `src/agent/loop_run/path_helpers.rs` to
host the workspace-relative path normalization + package.json/lock
sync helpers. Small follow-on extraction in the meta-issue #680
reduction effort.

Migrated to `path_helpers.rs` (`pub(super)`):

* `normalize_memory_path(raw_path, work_root) -> String` — projects
    a raw path token into a workspace-relative key for
    `WorkingMemory.touched_files`. SSOT used by 1 production caller
    in `turn.rs` + `verifier_orchestration.rs::record_repo_edit` +
    `scaffold_pipeline.rs::write_deterministic_scaffold_files`.
* `normalize_exploration_path(raw_path, work_root) -> String` —
    plan-mode exploration repeat-detector projection. Caller:
    `actor_loop_flow.rs::normalize_plan_exploration_key`.
* `sync_package_json_with_existing_lock(work_root, relative, content)
    -> String` — when a deterministic scaffold rewrites
    `package.json` in a worktree that already has a
    `package-lock.json`, replace dependency sections so the lock
    stays consistent. Caller:
    `scaffold_pipeline.rs::maybe_apply_deterministic_framework_support_files`.

`loop_run.rs` adjustments:

* Registered `mod path_helpers;` with a parent-#680 doc comment.

`turn.rs` adjustments:

* Removed the 3 fn definitions.
* Added `use super::path_helpers::normalize_memory_path;` for the
  production caller in `execute_tool_call` (records touched paths
  after a successful repo edit).
* Test mod call sites (`normalize_exploration_path` /
  `sync_package_json_with_existing_lock`) rewritten to
  `super::super::path_helpers::*`.
* Pruned the 2 names from the existing test-mod `super::{...}`
  re-export blocks.

`verifier_orchestration.rs` / `actor_loop_flow.rs` /
`scaffold_pipeline.rs` adjustments:

* Switched each from `super::turn::{...}`-group imports to dedicated
  `super::path_helpers::*` imports.

LOC delta:

* `turn.rs`: 27,905 → 27,827 (-78 LOC)
* `path_helpers.rs`: 0 → 132 (new)
* Cumulative `turn.rs`: 39,489 → 27,827 (-11,662 LOC, -29.5%)

`pub(super)` matching pre-extraction visibility. No facade re-export
(DR3-001).

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
